Refactoring code structure: adding contact list, contact item and chat for into core folder
