+++ src/app/components/+chat/chat.component.ts
@@ -3,8 +3,8 @@
  */
 
 import {Component} from '@angular/core';
-import {ChatFormComponent} from './chat-form/index';
-import {ContactListComponent} from './contact-list/index';
+import {ChatFormComponent} from '../core/chat/chat-form/index';
+import {ContactListComponent} from '../core/chat/contact-list/index';
 import {ContactModel} from '../models/contact-model/contact.model';
 
 @Component({
