+++ src/app/components/core/chat/chat-form/chat-form.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/variables";
+@import "../../../../../styles/variables";
 
 .top {
   width: 100%;
