+++ src/app/components/core/chat/contact-item/contact-item.component.html
