+++ src/app/components/core/chat/contact-item/contact-item.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/variables";
+@import "../../../../../styles/variables";
 
 .name {
   font-size: 14px;
