+++ src/app/components/core/chat/contact-item/contact-item.component.spec.ts
@@ -1,7 +1,7 @@
 /**
  * Test Component ChatContactItem
  */
-import {ContactItemComponent} from "./contact-item.component";
+import {ContactItemComponent} from "./contact-item.component.ts";
 import {Component} from "@angular/core";
 @Component({
     selector: 'test-contact-item',
