+++ src/app/components/core/chat/contact-item/contact-item.component.ts
@@ -3,14 +3,14 @@
  */
 
 import {Component, Output, Input, EventEmitter} from '@angular/core';
-import {ContactModel} from "../../models/contact-model/contact.model";
-import {MessageModel} from "../../models/message-model/message.model";
-import {ManageChatService} from "../../../shared/services/src/manage-chat.service";
+import {ContactModel} from "../../../models/contact-model/contact.model";
+import {MessageModel} from "../../../models/message-model/message.model";
+import {ManageChatService} from "../../../../shared/services/src/manage-chat.service";
 
 @Component({
     selector: 'contact-item',
     moduleId: module.id,
-    templateUrl: './contact-item.component.html',
+    templateUrl: 'contact-item.component.html',
     styleUrls : ['./contact-item.component.css']
 })
 export class ContactItemComponent {
