+++ src/app/components/core/chat/contact-item/index.ts
