+++ src/app/components/core/chat/contact-list/contact-list.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/variables";
+@import "../../../../../styles/variables";
 
 .contacts {
   margin-left: -1px;
