+++ src/app/components/core/chat/contact-list/index.ts
